backend/requirements.txt
# FastAPI 기본
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-multipart==0.0.6

# 설정 관리
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP 클라이언트 (외부 API 호출용)
httpx==0.25.2

# 기본 유틸리티
python-multipart==0.0.6

# 개발용 (선택사항)
pytest==7.4.3
pytest-asyncio==0.21.1

# TODO: 필요시 추가할 패키지들 (주석 처리)
# 데이터베이스
# sqlalchemy==2.0.23
# alembic==1.12.1

# 인증/보안
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4

# Redis (캐싱용)
# redis==5.0.1

# 이메일
# fastapi-mail==1.4.1

# 파일 업로드/처리
# python-magic==0.4.27
# pillow==10.1.0

#cv/데이터 (yolo+전처리)
numpy>=1.24
pandas>=2.1
pillow>=10.1

# 서버/컨테이너에선 아래를 권장(헤드리스):
# opencv-python-headless>=4.8.1.78
# 로컬 GUI가 필요하면 대신 이걸 사용:
opencv-python>=4.8.1.78

# YOLO 프레임워크
ultralytics>=8.3.0


